backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
aiofiles==23.2.1
pandas==2.1.3
numpy==1.26.2
jinja2==3.1.2
httpx==0.25.1
tenacity==8.2.3
loguru==0.7.2